requirements.txt
Flask==0.12.1
influxdb==4.1.0
pyaml==16.12.2